requirements.txt
gunicorn
gevent
gevent-websocket
flask
flask_socketio
yt_dlp[default]==2025.12.08
pyyaml
opencc-python-reimplemented
